feat(reader): migrate rail-segments logic to feature:reader commonMain (iOS parity) (#946)
First category-B slice of the Android↔iOS test-parity effort. The whole-book
chapter-rail generator and its CBZ/PDF siblings were pure Kotlin trapped in :app
only because they lived there; iOS could not share them.

- Move RailSegment, RailSegmentGenerator, CbzRailSegments, PdfRailSegments from
  :app (com.riffle.app.feature.reader) into feature:reader commonMain
  (com.riffle.feature.reader), shared with iOS. :app keeps typealias + forwarding
  shims (the AbsTargetResolver idiom) so no consumer changed.
- Promote core:domain normalizeEpubHref from jvmMain to commonMain, isolating the
  java.net.URI path extraction behind an internal `expect fun uriPath`: JVM actual
  is byte-identical (java.net.URI(raw).path); iOS actual uses NSURL. resolveEpubHref
  and epubCfiToSpineIndex come along and are now iOS-visible too.
- Move RailSegmentGeneratorTest (79), CbzRailSegmentsTest (13), PdfRailSegmentsTest
  (13), RailSegmentInvariantTest (1) into feature:reader commonTest — they now run on
  iOS + JVM (105 executing, 0 failures on each). JUnit -> kotlin.test; assert message
  args reordered to last.
- RailCorpusTest (2, golden real-scale corpus) stays in :app jvmTest: it loads JSON
  fixtures via javaClass.getResourceAsStream + kotlinx.serialization (host-only). It
  still pins the moved buildRailSegments through the :app shim.

Six RailSegmentGeneratorTest names contained "," or "()", which are illegal in
Kotlin/Native identifiers. They were renamed (punctuation removed) with behaviour
unchanged; the old names are declared below so the test guardrail records the rename.

Removed-test: three-level nesting — part expands to chapters, chapters keep and drop same-file section anchors
Removed-test: flat numeric-prefix run — a single run is NOT absorbed (could be real numbered chapters)
Removed-test: siblings at same level are evaluated independently — one Part expands, another keeps
Removed-test: Bulgarian-tales shape — tiny parent with tiny children NOT expanded (Детето съдия)
Removed-test: Bulgarian-tales shape — tiny parent with tiny children NOT expanded (Меден-Трета)
Removed-test: unmatched href falls back to preceding chapter by spine order, not chapter 0
