+++ app/src/main/kotlin/com/riffle/app/feature/reader/RailSegment.kt
@@ -1,8 +1,4 @@
 package com.riffle.app.feature.reader
 
-data class RailSegment(
-    val title: String,
-    val href: String,
-    val weight: Float = 1f,
-    val groupIndex: Int? = null,
-)
+// RailSegment has moved to feature:reader (commonMain, shared with iOS).
+typealias RailSegment = com.riffle.feature.reader.RailSegment
