+++ feature/reader/src/commonMain/kotlin/com/riffle/feature/reader/RailSegment.kt
@@ -1,0 +1,8 @@
+package com.riffle.feature.reader
+
+data class RailSegment(
+    val title: String,
+    val href: String,
+    val weight: Float = 1f,
+    val groupIndex: Int? = null,
+)
